requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
Jinja2==3.1.6
pydantic>=2.0.0
